refactor: extract synapse feedback, querying, and recent-queries modules from core
core.py (1811 lines) was the highest-risk file in the package. Three
clusters move to new stdlib-only modules of free functions taking the
NeuralMind instance, matching the existing pattern in server.py /
self_improve.py / memory.py:

- synapse_feedback.py — the synapse write-path: file co-activation,
  Edit/Write reuse detection, deleted-file decay, and query-time
  Hebbian reinforcement.
- recent_queries.py — the JSONL replay log, including the Windows
  msvcrt byte-0 / POSIX flock locking machinery, moved wholesale and
  unchanged.
- querying.py — read-side formatting: hybrid highlights, the file
  skeleton view, and the graph-view payload.

NeuralMind keeps thin delegating methods so the public surface and
every neuralmind.core patch point (is_memory_logging_enabled,
log_query_event, SynapseStore, NeuralMind) are unchanged; the consent
gate and _ensure_built() plumbing stay in core. RECENT_QUERIES_* stay
class attrs and flow into the moved functions as parameters, so
monkeypatched values still take effect. core.py is now 1220 lines.
